Separation of Functions: Flag Riddance
Getting rid of all flagged functions to improve code maintainability and security. Currently, both ADMIN and USER accounts are being modified based on flags, which is a major security risk.

Future Plan: Store logged in user as charArray. String is insecure. Currently not doing so since it is in a .csv environment.