Massive Update
Minor security update: char[] is used instead of password String. This prevents the plain text password to be stored as an object, however there may still be some hunting to get rid of all instances of processing the password as string. User and Admin flags being removed from program to follow best practices.